requirements.txt
bleach==6.2.0
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
kaggle==1.7.4.5
numpy==2.2.5
pandas==2.2.3
protobuf==6.31.0
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-slugify==8.0.4
pytz==2025.2
requests==2.32.3
setuptools==80.7.1
six==1.17.0
text-unidecode==1.3
tqdm==4.67.1
tzdata==2025.2
urllib3==2.4.0
webencodings==0.5.1
